requirements.txt
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
apache-airflow==2.9.3
pytest==8.2.2
pytest-mock==3.14.0
